migration_map: now dissipation is handled, and several contour throughout the integration time are done (100). Theses files are used by movie-growth to display changes in the disk while planets evolves. migration_map is now run only once. When 'total_torque.gnuplot' exists, the scripts do not run migration_map again. If changes have been made, you must run migration_map manually

+++ analysis/mercury-statistique.py
@@ -192,14 +192,15 @@ nb_simu = len(liste_simu)
 
 # If there is a disk, we get the torque diagram
 if isDisk:
-  shutil.copy2("%s/disk.in" % liste_simu[0], ".")
-  shutil.copy2("%s/param.in" % liste_simu[0], ".")
-  (process_stdout, process_stderr, returncode) = autiwa.lancer_commande(os.path.join(binaryPath, "migration_map"))
-  if (returncode != 0):
-    print(process_stdout)
-    print(process_stderr)
-    print("Process terminated with error %d" % returncode)
-    pdb.set_trace()
+  if not(os.path.isfile("total_torque.gnuplot")):
+    shutil.copy2("%s/disk.in" % liste_simu[0], ".")
+    shutil.copy2("%s/param.in" % liste_simu[0], ".")
+    (process_stdout, process_stderr, returncode) = autiwa.lancer_commande(os.path.join(binaryPath, "migration_map"))
+    if (returncode != 0):
+      print(process_stdout)
+      print(process_stderr)
+      print("Process terminated with error %d" % returncode)
+      pdb.set_trace()
     
   
   # We read the contour of the zero torque zones
